+++ MetasysServicesComExampleApp/FeaturesDemo/IComRefreshTokenDemo.cs
@@ -5,15 +5,21 @@ namespace MetasysServicesComExampleApp.FeaturesDemo
 {
     public class IComRefreshTokenDemo
     {
-        private LegacyMetasysClient legacyClient;
+        private ILegacyMetasysClient legacyClient;
 
-        public IComRefreshTokenDemo(LegacyMetasysClient legacyClient)
+        public IComRefreshTokenDemo(ILegacyMetasysClient legacyClient)
         {
             this.legacyClient = legacyClient;
         }
 
         public void Run()
         {
+            #region Refresh
+            Console.WriteLine("\nRefreshing Token...");
+            var token = legacyClient.Refresh();
+            Console.WriteLine($"Access token: {token.Token} expires {token.Expires}.");
+            Console.ReadLine();
+            #endregion
         }
     }
 }
